+++ frontend/src/components/DeploymentConfigTool/index.ts
@@ -1,7 +1,7 @@
 import DeploymentConfigTool from "./DeploymentConfigTool.vue";
 import DeploymentStage from "./DeploymentStage.vue";
-import DeploymentRoadmap from "./DeploymentRoadmap.vue";
+import DeploymentMatrix from "./DeploymentMatrix.vue";
 
 export default DeploymentConfigTool;
 
-export { DeploymentConfigTool, DeploymentStage, DeploymentRoadmap };
+export { DeploymentConfigTool, DeploymentStage, DeploymentMatrix };
